Rust 1.0.0 alpha changes, add `query_row_safe`

+++ rusqlite/src/types.rs
@@ -136,9 +136,15 @@ impl<T: ToSql> ToSql for Option<T> {
 /// ## Example
 ///
 /// ```rust,no_run
+/// #![allow(unstable)]
+/// # extern crate libc;
+/// # extern crate rusqlite;
 /// # use rusqlite::{SqliteConnection, SqliteResult};
 /// # use rusqlite::types::{Null};
-/// fn insert_null(conn: &SqliteConnection) -> SqliteResult<uint> {
+/// # use libc::{c_int};
+/// fn main() {
+/// }
+/// fn insert_null(conn: &SqliteConnection) -> SqliteResult<c_int> {
 ///     conn.execute("INSERT INTO people (name) VALUES (?)", &[&Null])
 /// }
 /// ```
@@ -186,7 +192,7 @@ impl FromSql for Vec<u8> {
         let c_blob = ffi::sqlite3_column_blob(stmt, col);
         let len = ffi::sqlite3_column_bytes(stmt, col);
 
-        assert!(len >= 0); let len = len as uint;
+        assert!(len >= 0); let len = len as usize;
 
         Ok(Vec::from_raw_buf(mem::transmute(c_blob), len))
     }
